feat: Connect CharacterEdit to Navigation and renamed it
